# chore: improve render string performance