--- a/render/text.go
+++ b/render/text.go
@@ -6,7 +6,6 @@ package render
 
 import (
 	"fmt"
-	"io"
 	"net/http"
 )
 
@@ -35,6 +34,6 @@ func WriteString(w http.ResponseWriter, format string, data []interface{}) (err 
 		_, err = fmt.Fprintf(w, format, data...)
 		return
 	}
-	_, err = io.WriteString(w, format)
+	_, err = w.Write([]byte(format))
 	return
 }